+++ src/rpc_api/data_types.rs
@@ -1,6 +1,7 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+use std::str::FromStr;
 use std::sync::Arc;
 
 use crate::beacon::BeaconSchedule;
@@ -29,19 +30,22 @@ use crate::state_manager::StateManager;
 use ahash::HashSet;
 use chrono::Utc;
 use cid::Cid;
+use fil_actor_interface::miner::MinerInfo;
 use fil_actor_interface::{
     market::{DealProposal, DealState},
     miner::MinerPower,
     power::Claim,
 };
+use fil_actor_miner_state::v12::{BeneficiaryTerm, PendingBeneficiaryChange};
 use fil_actors_shared::fvm_ipld_bitfield::BitField;
 use fvm_ipld_blockstore::Blockstore;
-use fvm_ipld_encoding::RawBytes;
+use fvm_ipld_encoding::{BytesDe, RawBytes};
 use jsonrpc_v2::{MapRouter as JsonRpcMapRouter, Server as JsonRpcServer};
 use libipld_core::ipld::Ipld;
+use libp2p::PeerId;
 use num_bigint::BigInt;
 use parking_lot::RwLock as SyncRwLock;
-use serde::{Deserialize, Serialize};
+use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use serde_json::Value;
 use tokio::sync::RwLock;
 
@@ -189,6 +193,206 @@ impl HasLotusJson for ApiMessage {
         ApiMessage {
             cid: lotus_json.cid.into_inner(),
             message: lotus_json.message.into_inner(),
+        }
+    }
+}
+
+const EMPTY_ADDRESS_VALUE: &str = "<empty>";
+
+/// This wrapper is needed because of a bug in Lotus.
+/// See: <https://github.com/filecoin-project/lotus/issues/11461>.
+#[derive(Debug, Eq, PartialEq, Clone)]
+pub struct AddressOrEmpty(pub Option<Address>);
+
+impl Serialize for AddressOrEmpty {
+    fn serialize<S>(&self, s: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let address_bytes = match self.0 {
+            Some(addr) => addr.to_string(),
+            None => EMPTY_ADDRESS_VALUE.to_string(),
+        };
+
+        s.collect_str(&address_bytes)
+    }
+}
+
+impl<'de> Deserialize<'de> for AddressOrEmpty {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let address_str = String::deserialize(deserializer)?;
+        if address_str.eq(EMPTY_ADDRESS_VALUE) {
+            return Ok(Self(None));
+        }
+
+        Address::from_str(&address_str)
+            .map_err(de::Error::custom)
+            .map(|addr| Self(Some(addr)))
+    }
+}
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct MinerInfoLotusJson {
+    #[serde(with = "crate::lotus_json")]
+    pub owner: Address,
+    #[serde(with = "crate::lotus_json")]
+    pub worker: Address,
+    pub new_worker: AddressOrEmpty,
+    #[serde(with = "crate::lotus_json")]
+    pub control_addresses: Vec<Address>, // Must all be ID addresses.
+    pub worker_change_epoch: ChainEpoch,
+    //#[serde(with = "crate::lotus_json")]
+    pub peer_id: Option<PeerId>,
+    #[serde(with = "crate::lotus_json")]
+    pub multiaddrs: Vec<Vec<u8>>,
+    pub window_po_st_proof_type: fvm_shared2::sector::RegisteredPoStProof,
+    pub sector_size: fvm_shared2::sector::SectorSize,
+    pub window_po_st_partition_sectors: u64,
+    pub consensus_fault_elapsed: ChainEpoch,
+    #[serde(with = "crate::lotus_json")]
+    pub pending_owner_address: Option<Address>,
+    #[serde(with = "crate::lotus_json")]
+    pub beneficiary: Address,
+    #[serde(with = "crate::lotus_json")]
+    pub beneficiary_term: BeneficiaryTerm,
+    #[serde(with = "crate::lotus_json")]
+    pub pending_beneficiary_term: Option<PendingBeneficiaryChange>,
+}
+
+impl HasLotusJson for MinerInfo {
+    type LotusJson = MinerInfoLotusJson;
+    fn snapshots() -> Vec<(serde_json::Value, Self)> {
+        vec![]
+    }
+    fn into_lotus_json(self) -> Self::LotusJson {
+        MinerInfoLotusJson {
+            owner: self.owner.into(),
+            worker: self.worker.into(),
+            new_worker: AddressOrEmpty(self.new_worker.map(|addr| addr.into())),
+            control_addresses: self
+                .control_addresses
+                .into_iter()
+                .map(|a| a.into())
+                .collect(),
+            worker_change_epoch: self.worker_change_epoch,
+            peer_id: PeerId::try_from(self.peer_id).ok(),
+            multiaddrs: self.multiaddrs.into_iter().map(|addr| addr.0).collect(),
+            window_po_st_proof_type: self.window_post_proof_type,
+            sector_size: self.sector_size,
+            window_po_st_partition_sectors: self.window_post_partition_sectors,
+            consensus_fault_elapsed: self.consensus_fault_elapsed,
+            // NOTE: In Lotus this field is never set for any of the versions, so we have to ignore
+            // it too.
+            // See: <https://github.com/filecoin-project/lotus/blob/b6a77dfafcf0110e95840fca15a775ed663836d8/chain/actors/builtin/miner/v12.go#L370>.
+            pending_owner_address: None,
+            beneficiary: self.beneficiary.into(),
+            beneficiary_term: self.beneficiary_term,
+            pending_beneficiary_term: self.pending_beneficiary_term,
+        }
+    }
+    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
+        MinerInfo {
+            owner: lotus_json.owner.into(),
+            worker: lotus_json.worker.into(),
+            new_worker: lotus_json.new_worker.0.map(|addr| addr.into()),
+            control_addresses: lotus_json
+                .control_addresses
+                .into_iter()
+                .map(|a| a.into())
+                .collect(),
+            worker_change_epoch: lotus_json.worker_change_epoch,
+            peer_id: lotus_json.peer_id.unwrap_or(PeerId::random()).to_bytes(),
+            multiaddrs: lotus_json.multiaddrs.into_iter().map(BytesDe).collect(),
+            window_post_proof_type: lotus_json.window_po_st_proof_type,
+            sector_size: lotus_json.sector_size,
+            window_post_partition_sectors: lotus_json.window_po_st_partition_sectors,
+            consensus_fault_elapsed: lotus_json.consensus_fault_elapsed,
+            // Ignore this field as it is never set on Lotus side.
+            pending_owner_address: None,
+            beneficiary: lotus_json.beneficiary.into(),
+            beneficiary_term: lotus_json.beneficiary_term,
+            pending_beneficiary_term: lotus_json.pending_beneficiary_term,
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct BeneficiaryTermLotusJson {
+    /// The total amount the current beneficiary can withdraw. Monotonic, but reset when beneficiary changes.
+    #[serde(with = "crate::lotus_json")]
+    pub quota: TokenAmount,
+    /// The amount of quota the current beneficiary has already withdrawn
+    #[serde(with = "crate::lotus_json")]
+    pub used_quota: TokenAmount,
+    /// The epoch at which the beneficiary's rights expire and revert to the owner
+    pub expiration: ChainEpoch,
+}
+
+impl HasLotusJson for BeneficiaryTerm {
+    type LotusJson = BeneficiaryTermLotusJson;
+
+    fn snapshots() -> Vec<(Value, Self)> {
+        vec![]
+    }
+
+    fn into_lotus_json(self) -> Self::LotusJson {
+        BeneficiaryTermLotusJson {
+            used_quota: self.used_quota.into(),
+            quota: self.quota.into(),
+            expiration: self.expiration,
+        }
+    }
+
+    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
+        Self {
+            used_quota: lotus_json.used_quota.into(),
+            quota: lotus_json.quota.into(),
+            expiration: lotus_json.expiration,
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct PendingBeneficiaryChangeLotusJson {
+    #[serde(with = "crate::lotus_json")]
+    pub new_beneficiary: Address,
+    #[serde(with = "crate::lotus_json")]
+    pub new_quota: TokenAmount,
+    pub new_expiration: ChainEpoch,
+    pub approved_by_beneficiary: bool,
+    pub approved_by_nominee: bool,
+}
+
+impl HasLotusJson for PendingBeneficiaryChange {
+    type LotusJson = PendingBeneficiaryChangeLotusJson;
+
+    fn snapshots() -> Vec<(Value, Self)> {
+        vec![]
+    }
+
+    fn into_lotus_json(self) -> Self::LotusJson {
+        PendingBeneficiaryChangeLotusJson {
+            new_beneficiary: self.new_beneficiary.into(),
+            new_quota: self.new_quota.into(),
+            new_expiration: self.new_expiration,
+            approved_by_beneficiary: self.approved_by_beneficiary,
+            approved_by_nominee: self.approved_by_nominee,
+        }
+    }
+
+    fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
+        Self {
+            new_beneficiary: lotus_json.new_beneficiary.into(),
+            new_quota: lotus_json.new_quota.into(),
+            new_expiration: lotus_json.new_expiration,
+            approved_by_beneficiary: lotus_json.approved_by_beneficiary,
+            approved_by_nominee: lotus_json.approved_by_nominee,
         }
     }
 }