+++ src/rpc_api/mod.rs
@@ -80,6 +80,7 @@ pub static ACCESS_MAP: Lazy<HashMap<&str, Access>> = Lazy::new(|| {
     access.insert(state_api::STATE_GET_ACTOR, Access::Read);
     access.insert(state_api::STATE_MARKET_BALANCE, Access::Read);
     access.insert(state_api::STATE_MARKET_DEALS, Access::Read);
+    access.insert(state_api::STATE_MINER_INFO, Access::Read);
     access.insert(state_api::STATE_MINER_ACTIVE_SECTORS, Access::Read);
     access.insert(state_api::STATE_MINER_FAULTS, Access::Read);
     access.insert(state_api::STATE_MINER_POWER, Access::Read);
@@ -237,6 +238,7 @@ pub mod state_api {
     pub const STATE_GET_ACTOR: &str = "Filecoin.StateGetActor";
     pub const STATE_MARKET_BALANCE: &str = "Filecoin.StateMarketBalance";
     pub const STATE_MARKET_DEALS: &str = "Filecoin.StateMarketDeals";
+    pub const STATE_MINER_INFO: &str = "Filecoin.StateMinerInfo";
     pub const STATE_MINER_FAULTS: &str = "Filecoin.StateMinerFaults";
     pub const STATE_MINER_POWER: &str = "Filecoin.StateMinerPower";
     pub const STATE_MINER_DEADLINES: &str = "Filecoin.StateMinerDeadlines";