+++ src/tool/subcommands/api_cmd.rs
@@ -393,6 +393,10 @@ fn snapshot_tests(store: &ManyCar, n_tipsets: usize) -> anyhow::Result<Vec<RpcTe
                     }
                 }
             }
+            tests.push(RpcTest::identity(ApiInfo::state_miner_info_req(
+                *block.miner_address(),
+                tipset.key().clone(),
+            )));
             tests.push(RpcTest::identity(ApiInfo::state_miner_power_req(
                 *block.miner_address(),
                 tipset.key().clone(),